refactor: split ClusterDetails page to smaller components
